FSA-5739: Renamed variables
