--- projects/schematics/src/add-fonts/schema.json
@@ -1,21 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema",
-  "id": "FontCopySchematics",
-  "title": "Font Copy Schematics",
-  "type": "object",
-  "properties": {
-    "project": {
-      "type": "string",
-      "description": "The name of the project.",
-      "$default": {
-        "$source": "projectName"
-      }
-    },
-    "name": {
-      "type": "string",
-      "description": "Copy font-icons to assets folder",
-      "default": "Srki"
-    }
-  },
-  "required": []
-}
